Revert naming scheme, add hub ID, don't throw on on/off press
Naming is tricky. The proposed naming scheme is okay for switches
that are actually connected to something in the Eaton hub, but for
switches that are exclusively used to generate events for HA it
does not work very well. These end up being named just "Rocker 1 ()"
which is not particularly useful.

Instead, including the "comp_name" will pull the name given the
switch in the Eaton app, making it easier to identify which
switch/rocker is which.