Various sensor fixes
Correctly assign SensorEntityDescription to the correct attribute.
This fixes long term statistics (state_class correctly recorded)
and removes the need for overriding some properties.

With the fixed SensorEntityDescription we have a malconfigured power
sensor however. Assuming the state class is supposed to be
measurement, the sensor device class probably needs to be POWER
with the unit being WATT.

For some reason, the energy sensor was constructed if the
temperature is available int he state value. Probably it should
depend on the power, not temperature (same as the power sensor).

The energy sensor also had two conflicting sensor states configured
(total and total increasing). Stuck with the latter for now,
removed the former.

Made the power calculation a bit more robust as well